+++ Plugins/Magic/src/main/java/com/tagnumelite/projecteintegration/plugins/magic/PluginExtraBotany.java
@@ -22,7 +22,6 @@
 package com.tagnumelite.projecteintegration.plugins.magic;
 
 import com.meteor.extrabotany.api.ExtraBotanyAPI;
-import com.meteor.extrabotany.common.crafting.recipe.RecipePedestal;
 import com.tagnumelite.projecteintegration.api.mappers.PEIMapper;
 import com.tagnumelite.projecteintegration.api.plugin.APEIPlugin;
 import com.tagnumelite.projecteintegration.api.plugin.PEIPlugin;
@@ -41,9 +40,7 @@ public class PluginExtraBotany extends APEIPlugin {
 
         @Override
         public void setup() {
-            for (RecipePedestal recipe : ExtraBotanyAPI.pedestalRecipes) {
-                addRecipe(recipe.getOutput(), recipe.getInput());
-            }
+            ExtraBotanyAPI.pedestalRecipes.forEach(r -> addRecipe(r.getOutput(), r.getInput()));
         }
     }
 }