+++ Plugins/Magic/src/main/java/com/tagnumelite/projecteintegration/plugins/magic/PluginPsi.java
@@ -25,12 +25,11 @@ import com.tagnumelite.projecteintegration.api.mappers.PEIMapper;
 import com.tagnumelite.projecteintegration.api.plugin.APEIPlugin;
 import com.tagnumelite.projecteintegration.api.plugin.PEIPlugin;
 import vazkii.psi.api.PsiAPI;
-import vazkii.psi.api.recipe.TrickRecipe;
 
 @PEIPlugin("psi")
 public class PluginPsi extends APEIPlugin {
     @Override
-    public void setup() throws Exception {
+    public void setup() {
         addMapper(new TrickMapper());
     }
 
@@ -41,9 +40,7 @@ public class PluginPsi extends APEIPlugin {
 
         @Override
         public void setup() {
-            for (TrickRecipe recipe : PsiAPI.trickRecipes) {
-                addRecipe(recipe.getOutput(), recipe.getInput());
-            }
+            PsiAPI.trickRecipes.forEach(r -> addRecipe(r.getOutput(), r.getInput()));
         }
     }
 }