+++ Plugins/Tech/src/main/java/com/tagnumelite/projecteintegration/plugins/tech/PluginCompactMachines.java
@@ -24,8 +24,6 @@ package com.tagnumelite.projecteintegration.plugins.tech;
 import com.tagnumelite.projecteintegration.api.mappers.PEIMapper;
 import com.tagnumelite.projecteintegration.api.plugin.APEIPlugin;
 import com.tagnumelite.projecteintegration.api.plugin.PEIPlugin;
-import net.minecraft.item.ItemStack;
-import org.dave.compactmachines3.miniaturization.MultiblockRecipe;
 import org.dave.compactmachines3.miniaturization.MultiblockRecipes;
 
 @PEIPlugin("compactmachines3")
@@ -42,10 +40,7 @@ public class PluginCompactMachines extends APEIPlugin {
 
         @Override
         public void setup() {
-            for (MultiblockRecipe recipe : MultiblockRecipes.getRecipes()) {
-                addRecipe(recipe.getTargetStack(), recipe.getCatalystStack(),
-                    recipe.getRequiredItemStacks().toArray(new ItemStack[0]));
-            }
+            MultiblockRecipes.getRecipes().forEach(r -> addRecipe(r.getTargetStack(), r.getCatalystStack(), r.getRequiredItemStacks()));
         }
     }
 }