+++ Plugins/Tech/src/main/java/com/tagnumelite/projecteintegration/plugins/tech/PluginDraconicEvolution.java
@@ -23,19 +23,15 @@ package com.tagnumelite.projecteintegration.plugins.tech;
 
 import com.brandon3055.draconicevolution.DEFeatures;
 import com.brandon3055.draconicevolution.api.fusioncrafting.FusionRecipeAPI;
-import com.brandon3055.draconicevolution.api.fusioncrafting.IFusionRecipe;
 import com.tagnumelite.projecteintegration.api.mappers.PEIMapper;
 import com.tagnumelite.projecteintegration.api.plugin.APEIPlugin;
 import com.tagnumelite.projecteintegration.api.plugin.PEIPlugin;
-import net.minecraft.item.ItemStack;
 
 @PEIPlugin("draconicevolution")
 public class PluginDraconicEvolution extends APEIPlugin {
     @Override
     public void setup() {
         addEMC(DEFeatures.draconiumDust, 8192);
-        // TODO: Figure out why ItemPersitant crashes the plugin when Draconic Evolution isn't installed
-        // Raises classNotDefError
         addEMC(DEFeatures.dragonHeart, 262144);
         addEMC(DEFeatures.chaosShard, 1024000);
 
@@ -49,9 +45,7 @@ public class PluginDraconicEvolution extends APEIPlugin {
 
         @Override
         public void setup() {
-            for (IFusionRecipe recipe : FusionRecipeAPI.getRecipes()) {
-                addRecipe(recipe.getRecipeOutput(ItemStack.EMPTY), recipe.getRecipeIngredients().toArray());
-            }
+            FusionRecipeAPI.getRecipes().forEach(r -> addRecipe(r.getRecipeOutput(null), r.getRecipeIngredients()));
         }
     }
 }